+++ app/services/agent/supervisor.py
@@ -12,18 +12,13 @@ import json
 import logging
 import yaml
 
-from datetime import datetime
-from collections.abc import Awaitable, Callable
-from typing import cast
 from langchain.agents import create_agent
 from langchain_core.language_models.chat_models import BaseChatModel
 from langchain_core.runnables.config import RunnableConfig, ensure_config
 from langchain_core.tools import BaseTool, StructuredTool
 from langgraph.graph.state import Any, CompiledStateGraph, Checkpointer
-from langchain.agents.middleware import wrap_tool_call, SummarizationMiddleware
-from langchain.agents.middleware.types import AgentMiddleware
+from langchain.agents.middleware import SummarizationMiddleware
 from langchain.messages import AIMessage, ToolMessage
-from langchain.tools.tool_node import ToolCallRequest
 import langgraph.types
 from langgraph.types import Command
 from langchain_core.callbacks.manager import dispatch_custom_event
@@ -32,20 +27,14 @@ from .loader import AgentConfig
 from .system_prompts import SUPERVISOR_PROMPT
 from .middleware import (
     INTERRUPT_CANCEL_MESSAGE,
+    ChildAgentCancelled,
     MessagesHistoryMiddleware,
     create_cancel_check_middleware,
     inject_additional_kwargs_middleware,
     create_ui_tools_middleware,
+    _create_child_agent_middleware,
 )
 
-
-class ChildAgentCancelled(Exception):
-    """Raised when a child agent's tool execution is cancelled by the user."""
-
-    def __init__(self, agent_name: str, interrupt_info: dict | None = None):
-        super().__init__(agent_name)
-        self.agent_name = agent_name
-        self.interrupt_info = interrupt_info
 
 @dataclass
 class ChildAgent:
@@ -126,7 +115,7 @@ def create_supervisor_agent(
             create_cancel_check_middleware(),
             inject_additional_kwargs_middleware(),
             create_ui_tools_middleware(llm),
-            SummarizationMiddleware(model=llm, trigger=[("messages", 4), ("tokens", 6000)], keep=("messages", 4)),
+            SummarizationMiddleware(model=llm, trigger=[("messages", 20), ("tokens", 20000)], keep=("messages", 4)),
         ],
     )
 
@@ -309,85 +298,9 @@ def _create_agent_tool(child_agent: ChildAgent) -> BaseTool:
         response_format="content_and_artifact",
     )
 
-# =============================================================================
-# Middleware
-# =============================================================================
-
-
 def _dispatch_subagent_event(tag: str, name: str, query: str | None = None) -> None:
     """Dispatch a subagent lifecycle event with a valid JSON payload."""
     if query is not None:
         payload: dict = {"name": name}
         payload["query"] = query
         dispatch_custom_event("subagent_call", f"<{tag}>{json.dumps(payload)}</{tag}>")
-
-
-def _create_child_agent_middleware() -> AgentMiddleware:
-    """
-    Build the wrap-tool-call middleware that intercepts every child-agent tool call
-    made by the supervisor.  It has two responsibilities:
-
-    1. **Artifact propagation** — child agents return ``(str, dict)`` with a rich
-       artifact (mcp_responses, mcp_data, interrupt_info).  LangGraph's ToolNode
-       stores the artifact on the ToolMessage but does *not* copy it into
-       ``additional_kwargs``, so downstream code that reads ``additional_kwargs``
-       would miss it.  This middleware does that copy after each successful call.
-
-    2. **Cancellation short-circuit** — if the user rejected a human-in-the-loop
-       prompt inside a child agent, ``_invoke`` raises ``ChildAgentCancelled``.
-       The middleware catches it, constructs a well-formed ToolMessage, and returns
-       ``Command(goto="__end__")`` to terminate the supervisor graph immediately
-       without invoking the LLM node again.
-    """
-
-    @wrap_tool_call  # type: ignore[misc]
-    async def handle_child_agent_tool_call(
-        request: ToolCallRequest,
-        handler: Callable[[ToolCallRequest], Awaitable[ToolMessage | Command]],
-    ) -> ToolMessage | Command:
-        """Intercept a single child-agent tool call: propagate artifacts and handle cancellation."""
-        name = request.tool_call["name"]
-        query = request.tool_call["args"].get("query")
-        try:
-            logging.debug(f"Supervisor is invoking tool '{name}'")
-            result = await handler(request)
-
-            # Copy artifact fields (mcp_response, mcp_data, interrupt_info) into
-            # additional_kwargs so that consumers reading the ToolMessage directly
-            # (e.g. streaming handlers, memory middleware) can access them.
-            artifact = getattr(result, 'artifact', None) or {}
-            if artifact and isinstance(result, ToolMessage):
-                extra = {}
-                mcp_responses = artifact.get("mcp_responses", [])
-                if mcp_responses:
-                    extra["mcp_response"] = "\n".join(mcp_responses)
-                mcp_data = artifact.get("mcp_data")
-                if mcp_data:
-                    extra["mcp_data"] = mcp_data
-                interrupt_info = artifact.get("interrupt_info")
-                if interrupt_info:
-                    extra.update(interrupt_info)
-                    extra["created_at"] = datetime.now().isoformat()
-                if extra:
-                    result.additional_kwargs = {
-                        **result.additional_kwargs,
-                        **extra,
-                    }
-
-            return result
-        except ChildAgentCancelled as exc:
-            # Create a proper ToolMessage so the supervisor's state stays clean,
-            # then use Command to route directly to END — skipping the LLM call.
-            additional_kwargs = exc.interrupt_info or {}
-            additional_kwargs["created_at"] = datetime.now().isoformat()
-            tool_message = ToolMessage(
-                content=INTERRUPT_CANCEL_MESSAGE,
-                name=name,
-                tool_call_id=request.tool_call["id"],
-                additional_kwargs=additional_kwargs,
-            )
-            return Command(goto="__end__", update={"messages": [tool_message]})
-        except Exception as e:
-            raise
-
-    return cast(AgentMiddleware, handle_child_agent_tool_call)
