+++ tests/integration/test_single_agent.py
@@ -1,6 +1,7 @@
 from fastapi.testclient import TestClient
 from app.main import app
 from app.services.agent.loader import RANCHER_AGENT_PROMPT, AgentConfig, AuthenticationType
+from app.services.agent.child import CHILD_TOOL_USE_INSTRUCTIONS
 from app.services.agent.system_prompts import IDENTITY_PREAMBLE
 from app.services.llm import LLMManager
 from app.services.memory import StorageType
@@ -115,7 +116,7 @@ def test_websocket_single_prompt():
         assert messages == expected_messages_send_to_websocket
         assert len(fake_llm.all_calls) == 1, "Expected 1 LLM call"
         assert fake_llm.all_calls[0] == [
-            SystemMessage(content=RANCHER_AGENT_PROMPT),
+            SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS),
             HumanMessage(content="fake prompt"),
             SystemMessage(content=IDENTITY_PREAMBLE),
         ], "First call should have system prompt and user message"
@@ -154,12 +155,12 @@ def test_websocket_multiple_prompts():
         assert messages == expected_messages_send_to_websocket
         assert len(fake_llm.all_calls) == 2, "Expected 2 LLM calls"
         assert fake_llm.all_calls[0] == [
-            SystemMessage(content=RANCHER_AGENT_PROMPT),
+            SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS),
             HumanMessage(content="fake prompt 1"),
             SystemMessage(content=IDENTITY_PREAMBLE),
         ], "First call should have system prompt and first user message"
         assert fake_llm.all_calls[1] == [
-            SystemMessage(content=RANCHER_AGENT_PROMPT),
+            SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS),
             HumanMessage(content="fake prompt 1"),
             SystemMessage(content=IDENTITY_PREAMBLE),
             AIMessage(content="fake llm response 1"),
@@ -204,13 +205,13 @@ def test_websocket_tool_call():
         assert messages == expected_messages_send_to_websocket
         assert len(fake_llm.all_calls) == 2, "Expected 2 LLM calls (initial + after tool)"
         assert fake_llm.all_calls[0] == [
-            SystemMessage(content=RANCHER_AGENT_PROMPT),
+            SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS),
             HumanMessage(content="sum 4 + 5"),
             SystemMessage(content=IDENTITY_PREAMBLE),
         ], "First call should have system prompt and user message"
         # Second call includes tool call and result
         second_call = fake_llm.all_calls[1]
-        assert second_call[0] == SystemMessage(content=RANCHER_AGENT_PROMPT), "Second call should have system prompt"
+        assert second_call[0] == SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS), "Second call should have system prompt"
         assert second_call[1] == HumanMessage(content="sum 4 + 5"), "Second call should have user message"
         assert second_call[2] == SystemMessage(content=IDENTITY_PREAMBLE), "Second call should have identity preamble"
         assert isinstance(second_call[3], AIMessage) and second_call[3].tool_calls[0]["name"] == "add", "Second call should have AI message with tool call"
@@ -271,14 +272,14 @@ def test_conversation_history():
         
         # First call - just prompt 1
         assert fake_llm.all_calls[0] == [
-            SystemMessage(content=RANCHER_AGENT_PROMPT),
+            SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS),
             HumanMessage(content=fake_prompt_1),
             SystemMessage(content=IDENTITY_PREAMBLE),
         ], "First call should have system prompt and first user message"
         
         # Second call - prompt 1 + response 1 + prompt 2
         assert fake_llm.all_calls[1] == [
-            SystemMessage(content=RANCHER_AGENT_PROMPT),
+            SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS),
             HumanMessage(content=fake_prompt_1),
             SystemMessage(content=IDENTITY_PREAMBLE),
             AIMessage(content=fake_llm_response_1),
@@ -287,7 +288,7 @@ def test_conversation_history():
         
         # Third call - full history up to prompt 3
         assert fake_llm.all_calls[2] == [
-            SystemMessage(content=RANCHER_AGENT_PROMPT),
+            SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS),
             HumanMessage(content=fake_prompt_1),
             SystemMessage(content=IDENTITY_PREAMBLE),
             AIMessage(content=fake_llm_response_1),
@@ -298,7 +299,7 @@ def test_conversation_history():
         
         # Fourth call - full history up to prompt 4
         assert fake_llm.all_calls[3] == [
-            SystemMessage(content=RANCHER_AGENT_PROMPT),
+            SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS),
             HumanMessage(content=fake_prompt_1),
             SystemMessage(content=IDENTITY_PREAMBLE),
             AIMessage(content=fake_llm_response_1),
@@ -311,7 +312,7 @@ def test_conversation_history():
         
         # Fifth call - full history up to prompt 5
         assert fake_llm.all_calls[4] == [
-            SystemMessage(content=RANCHER_AGENT_PROMPT),
+            SystemMessage(content=RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS),
             HumanMessage(content=fake_prompt_1),
             SystemMessage(content=IDENTITY_PREAMBLE),
             AIMessage(content=fake_llm_response_1),
@@ -374,7 +375,7 @@ def test_websocket_with_ui_tools():
             mock_load.return_value = [agent_config]
             
             # Patch load_ui_tools_from_configmap to return our UI tools config
-            with patch('app.services.agent.middleware.load_ui_tools_from_configmap') as mock_load_configmap:
+            with patch('app.services.agent.middleware.ui_tools.load_ui_tools_from_configmap') as mock_load_configmap:
                 # Setup to return the UI tool and config directly
                 ui_tools_config = UIToolsConfigData(
                     tools=[ui_tool],
@@ -383,7 +384,7 @@ def test_websocket_with_ui_tools():
                 mock_load_configmap.return_value = ui_tools_config
                 
                 # Mock the UI tools selector
-                with patch('app.services.agent.middleware.create_ui_tools_selector') as mock_selector_factory:
+                with patch('app.services.agent.middleware.ui_tools.create_ui_tools_selector') as mock_selector_factory:
                     mock_selector = MagicMock()
                     mock_selector_factory.return_value = mock_selector
                     # Mock select_tools to return a formatted UI tool
@@ -421,7 +422,7 @@ def test_websocket_with_ui_tools():
                     # Verify workflow correctness: response content in stream
                     assert "Here is the resource information" in full_stream, "Response should contain LLM response"
                     assert len(fake_llm.all_calls) == 1, "Expected 1 LLM call"
-                    assert fake_llm.all_calls[0][0].content == RANCHER_AGENT_PROMPT, "LLM should receive system prompt"
+                    assert fake_llm.all_calls[0][0].content == RANCHER_AGENT_PROMPT + CHILD_TOOL_USE_INSTRUCTIONS, "LLM should receive system prompt"
                     assert "show me the resource" in fake_llm.all_calls[0][1].content, "LLM should receive user prompt"
                     assert fake_llm.all_calls[0][2].content == IDENTITY_PREAMBLE, "LLM should receive identity preamble"
