+++ tests/unit/services/agent/test_supervisor_agent.py
@@ -19,8 +19,8 @@ from app.services.agent.supervisor import (
     _build_child_config,
     _extract_last_message,
     _create_agent_tool,
-    _create_child_agent_middleware,
 )
+from app.services.agent.middleware import _create_child_agent_middleware
 from app.services.agent.loader import AgentConfig, AuthenticationType
 
 
@@ -543,7 +543,7 @@ class TestCreateSubagentEventMiddleware:
     @pytest.fixture
     def middleware_fn(self):
         """Inner monitor_tool with wrap_tool_call bypassed (identity decorator)."""
-        with patch("app.services.agent.supervisor.wrap_tool_call", lambda f: f):
+        with patch("app.services.agent.middleware.child_agent_tool.wrap_tool_call", lambda f: f):
             fn = _create_child_agent_middleware()
         return fn
